requirements.txt
requests
beautifulsoup4
lxml
pandas
schedule
python-dotenv
yagmail
